main.c moved to src/
